Fix: 새로운 계정으로 로그인 시 이전 계정의 좋아요 영화 데이터 유지현상 해결, 계정 삭제 기능 추가
가입 및 로그인 시 유저 객체를 UserDefaults에 저장 후 LibraryVM의  @Published var favoriteMovies 변수에 해당 유저의 아이디를 기반으로 DocumentDirecotry에 저장된 데이터를 할당.
AuthenicationView의 하위 뷰에서도 LibraryVM에 접근할 수 있도록 App의 시작점에서 해당 뷰모델을 생성하고 environment에 주입
SignIn, SignUp Button 분리

+++ Movie/ViewModels/Authentication/SettingsViewModel.swift
@@ -13,6 +13,9 @@ final class SettingsViewModel: ObservableObject {
         try AuthenticationManager.shared.signOut()
     }
     func deleteAccount() async throws {
+        guard let user = UserDefaults.standard.loadUser(DBUser.self, forKey: .user) else { return }
         try await AuthenticationManager.shared.delete()
+        try await UserManager.shared.deleteUser(userId: user.userId)
+        UserDefaults.standard.removeObject(forKey: "user")
     }
 }
